Refactor assistant progress toast into own component
