+++ src/leapfrogai_ui/src/lib/components/LFToast.svelte
@@ -1,86 +1,25 @@
 <script lang="ts">
   import { P } from 'flowbite-svelte';
-  import {
-    BanOutline,
-    CheckCircleSolid,
-    ExclamationCircleOutline,
-    InfoCircleOutline
-  } from 'flowbite-svelte-icons';
-  import type { ToastKind, ToastNotificationProps } from '$lib/types/toast';
-  import AssistantProgressToastContent from '$components/AssistantProgressToastContent.svelte';
+  import type { ToastNotificationProps } from '$lib/types/toast';
   import ToastOverride from '$components/ToastOverride.svelte';
+  import { getColor, getIconComponent } from '$helpers/toastHelpers';
 
   export let toast: ToastNotificationProps;
 
-  let { id, title, subtitle, kind, variant } = toast;
+  let { title, subtitle, kind } = toast;
 
   $: color = getColor(kind);
-
-  function getColor(toastKind: ToastKind) {
-    switch (toastKind) {
-      case 'success':
-        return 'green';
-      case 'info':
-        return 'blue';
-      case 'warning':
-        return 'yellow';
-      case 'error':
-        return 'red';
-      default:
-        return 'blue';
-    }
-  }
-
-  function getIconComponent(toastKind: ToastKind) {
-    switch (toastKind) {
-      case 'success':
-        return CheckCircleSolid;
-      case 'info':
-        return InfoCircleOutline;
-      case 'warning':
-        return ExclamationCircleOutline;
-      case 'error':
-        return BanOutline;
-      default:
-        return InfoCircleOutline;
-    }
-  }
-
-  function getAssistantVariantTitle(toastKind: ToastKind) {
-    switch (toastKind) {
-      case 'success':
-        return 'Assistant Updated';
-      case 'info':
-        return 'Updating Assistant Files';
-      case 'warning':
-        return 'Updating Assistant Files';
-      case 'error':
-        return 'Error Updating Assistant';
-      default:
-        return 'Updating Assistant Files';
-    }
-  }
 </script>
 
-<ToastOverride {color} align={false}>
+<ToastOverride {color}>
   <svelte:fragment slot="icon">
     <svelte:component this={getIconComponent(kind)} class="h-5 w-5" />
     <span class="sr-only">Toast icon</span>
   </svelte:fragment>
   <div class="flex flex-col">
-    {variant === 'assistant-progress' ? getAssistantVariantTitle(kind) : title}
+    {title}
     {#if subtitle}
       <P size="xs">{subtitle}</P>
-    {/if}
-    {#if variant === 'assistant-progress'}
-      <AssistantProgressToastContent
-        toastId={id}
-        fileIds={toast.fileIds}
-        vectorStoreId={toast.vectorStoreId}
-        on:statusChange={(e) => {
-          kind = e.detail;
-        }}
-      />
     {/if}
     <P size="xs">{`Time stamp [${new Date().toLocaleTimeString()}]`}</P>
   </div>
